project/requirements.txt

asgiref==3.3.1
boto3==1.17.29
botocore==1.20.29
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-ckeditor==6.0.0
django-ckeditor-5==0.0.12
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-discover-runner==1.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-ipware==3.0.2
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
django-role-permissions==3.1.1
django-rosetta==0.9.3
django-storages==1.11.1
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-recaptcha==2.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
mysqlclient==2.0.3
phonenumbers==8.12.20
Pillow==8.1.0
polib==1.1.0
protobuf==3.15.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
twilio==6.55.0
uritemplate==3.0.1
urllib3==1.26.3

